% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind.modello_number}
\alias{bind.modello_number}
\alias{bind.list}
\alias{bind}
\title{Bind}
\usage{
\method{bind}{modello_number}(..., k)

\method{bind}{list}(x, k, ...)

bind(x, ...)
}
\arguments{
\item{...}{\code{numbers} to bind together}

\item{k}{dimesion along to perform the bind}

\item{x}{list of \code{numbers} to bind together}
}
\value{
Returns a \code{number}
}
\description{
S3 method to bind \code{numbers} together
along a certain dimension.
}
\examples{
modello.init()
x1 = number(c(1, 2, 3))
y1 = bind(x1, x1, k=1)
print(y1)
print(y1$v)
x2 = number(matrix(c(1, 2, 3), 3, 1))
y2 = bind(x2, x2, k=2)
print(y2)
print(y2$v)
modello.close()
modello.init()
x1 = number(c(1, 2, 3))
y1 = bind(list(x1, x1), k=1)
print(y1)
print(y1$v)
x2 = number(matrix(c(1, 2, 3), 3, 1))
y2 = bind(list(x2, x2), k=2)
print(y2)
print(y2$v)
modello.close()
}
\author{
Filippo Monari
}
