% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim.modello_number}
\alias{dim.modello_number}
\title{Get Number Shape}
\usage{
\method{dim}{modello_number}(x)
}
\arguments{
\item{x}{a reference object of class 'number'}
}
\value{
Returns the shape vector of the \code{number}
}
\description{
S3 method to get the shape of a \code{number}
}
\examples{
modello.init(10, 10, 10, 10)
x = number(matrix(rnorm(9), 3, 3))
dim(x) # c(3, 3)
modello.close()
}
\author{
Filippo Monari
}
