% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.ids}
\alias{.ids}
\title{Get IDs}
\usage{
.ids(x)
}
\arguments{
\item{x}{list of reference objects}
}
\value{
Return a vector with the ids
}
\description{
Helper function to get the ids from a list of
reference object
}
\author{
Filippo Monari
}
