% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.k}
\alias{.k}
\title{Create constnat scalar number}
\usage{
.k(x)
}
\arguments{
\item{x}{value of the constant}
}
\value{
Returns a reference object of class 'number'
}
\description{
Create a constant scalar \code{number}
}
\examples{
modello.init()
x = .k(1)
print(x)
print(x$v)
modello.close()
}
\author{
Filippo Monari
}
