% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\docType{data}
\name{.modello}
\alias{.modello}
\title{Object containing the corrent session.}
\format{
An object of class \code{modello_session} (inherits from \code{R6}) of length 40.
}
\usage{
.modello
}
\description{
Object containing the corrent session.
}
\keyword{datasets}
