% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{embeddings}
\alias{embeddings}
\title{Embeddings}
\usage{
embeddings(xf, xx, n)
}
\arguments{
\item{xf}{a reference object of class 'number' referring to the factors}

\item{xx}{a reference object of class 'number' referring to the embedding vectors}

\item{n}{number of embedding vector to bind together}
}
\value{
Returns a reference object of class 'number'
}
\description{
Embeddings
}
\author{
Filippo Monari
}
