% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_math.R
\name{exp.modello_number}
\alias{exp.modello_number}
\title{Number Exponential}
\usage{
\method{exp}{modello_number}(x)
}
\arguments{
\item{x}{a reference object of class 'number'}
}
\value{
Returns a reference object of class 'number'
}
\description{
Calculates the exponential of a \code{number}
}
\examples{
modello.init()
## For sclars
x1 = number(-3, scalar=TRUE)
y1 = exp(x1)
print(y1)
print(y1$v)
## For arrays
x2 = number(matrix(-3, 3, 3))
y2 = exp(x2)
print(y2)
print(y2$v)
modello.close()
}
\author{
Filippo Monari
}
