% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_math.R
\name{\%.T\%}
\alias{\%.T\%}
\title{Matrix - Matrix Traspose Multiplication}
\usage{
A \%.T\% B
}
\arguments{
\item{A}{a reference object of class 'number' with rank 2}

\item{B}{a reference object of class 'number' with rank 2}
}
\value{
Returns a reference object of class 'number'
}
\description{
Calculates:
ans = A.B**T 
where A and B are matrices
}
\examples{
modello.init()
A = number(matrix(rnorm(9), 3, 3))
B = number(matrix(rnorm(9), 3, 3))
b = number(matrix(rnorm(3), 1, 3))
ans1 = A \%.T\% B
ans2 = A \%.T\% b
print(ans1)
print(ans1$v)
print(ans2)
print(ans2$v)
modello.close()
}
\author{
Filippo Monari
}
