% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_math.R
\name{\%T.T\%}
\alias{\%T.T\%}
\title{Matrix Traspose - Matrix Traspose Multiplication}
\usage{
A \%T.T\% B
}
\arguments{
\item{A}{a reference object of class 'number' with rank 2}

\item{B}{a reference object of class 'number' with rank 2}
}
\value{
Returns a reference object of class 'number'
}
\description{
Calculates:
ans = A**T.B**T 
where A and B are matrices
}
\examples{
modello.init()
A = number(matrix(rnorm(9), 3, 3))
B = number(matrix(rnorm(9), 3, 3))
ans = A \%T.T\% B
print(ans)
print(ans$v)
modello.close()
}
\author{
Filippo Monari
}
