% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{graph.close}
\alias{graph.close}
\title{Close Graph}
\usage{
graph.close()
}
\value{
Returns invisible NULL
}
\description{
Closes the last open \code{graph}
}
\details{
Only one open \code{graph} at a time is allowed.
}
\examples{
modello.init()
x1 = number(1)
x2 = number(3)
x3 = number(2)
g = graph.open()
y = log((x1 + x3)**x2)
graph.close()
print(y$v)
x1$v = 2
g$op()
print(y$v)
y$dv = 1
g$bw()
print(x1$dv)
print(x2$dv)
print(x3$dv)
modello.close()
}
\author{
Filippo Monari
}
