% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_math.R
\name{invMat.modello_number}
\alias{invMat.modello_number}
\alias{invMat.matrix}
\alias{invMat}
\title{Number Inverse Matrix}
\usage{
\method{invMat}{modello_number}(A)

\method{invMat}{matrix}(A)

invMat(A)
}
\arguments{
\item{A}{a 'numeric' or a reference object of class 'number'}
}
\value{
Returns a 'numeric' or reference object of class 'number'
}
\description{
Calculates the inverse matrix
}
\examples{
modello.init()
## For modello_number matrices
x = number(diag(0.1, 3))
ix = invMat(x)
print(ix)
print(ix$v)
modello.close()
## For matrices
x = diag(0.1, 3)
ix = invMat(x)
print(ix)
}
\author{
Filippo Monari
}
