% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.graph}
\alias{is.graph}
\title{Is Graph}
\usage{
is.graph(g)
}
\arguments{
\item{g}{object}
}
\value{
Returns TRUE/FALSE
}
\description{
Checks if an object os of class 'graph'
}
\examples{
modello.init()
g = graph.open()
is.graph(g) # TRUE
modello.close()
}
\author{
Filippo Monari
}
