% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.number}
\alias{is.number}
\title{Is Number}
\usage{
is.number(x)
}
\arguments{
\item{x}{an object}
}
\value{
Returns TRUE/FALSE
}
\description{
Checks if a object is of class 'number'
}
\examples{
modello.init()
x = number(1)
is.number(x) # TRUE
modello.close()
}
\author{
Filippo Monari
}
