% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.opt}
\alias{is.opt}
\title{Is Optimiser}
\usage{
is.opt(x)
}
\arguments{
\item{x}{objcet}
}
\value{
Return TRUE/FALSE
}
\description{
Check if an object is an optimiser.
}
\examples{
modello.init(10, 10, 10, 10)
mdl = module.lm$new(1, 3, 1)
optimizer = adam.opt(mdl$pars())
is.opt(optimizer) # TRUE
modello.close()
}
\author{
Filippo Monari
}
