% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{length.modello_number}
\alias{length.modello_number}
\title{Number Length}
\usage{
\method{length}{modello_number}(x)
}
\arguments{
\item{x}{a referecen object of class 'number'}
}
\value{
Returns the size of the \code{number}
}
\description{
S3 method to get the total size of a \code{number}
}
\examples{
modello.init(10, 10, 10, 10)
x = number(rnorm(3))
length(x) #3
modello.close()
}
\author{
Filippo Monari
}
