% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_math.R
\name{log.modello_number}
\alias{log.modello_number}
\title{Number Log}
\usage{
\method{log}{modello_number}(x, base = NULL)
}
\arguments{
\item{x}{a reference object of class 'number'}

\item{base}{ignored, for compatibility with the generic function}
}
\value{
Returns a reference object of class 'number'
}
\description{
Calculates the logarithm of a \code{number}
}
\examples{
modello.init(10, 10, 10, 10)
## For sclars
x1 = number(3, scalar=TRUE)
y1 = log(x1)
print(y1)
print(y1$v)
## For arrays
x2 = number(matrix(3, 3, 3))
y2 = log(x2)
print(y2)
print(y2$v)
modello.close()
}
\author{
Filippo Monari
}
