% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modello.close}
\alias{modello.close}
\title{Close Modello Session}
\usage{
modello.close()
}
\value{
Returns invsible NULL
}
\description{
Deallocates all the arrays.
}
\examples{
modello.init(10, 10, 10, 10) # Initilaises the modello session with arrays of the given size
modello.is.init() # TRUE
modello.reset()   # restart the modello session with initial array sizes 
modello.is.init() # TRUE
modello.close()   # Closes modello
modello.is.init() # FALSE
}
\author{
Filippo Monari
}
