% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modello.gc}
\alias{modello.gc}
\title{Modello Session Garbage Collector}
\usage{
modello.gc(env = .GlobalEnv)
}
\arguments{
\item{env}{environmet wherein collect the garbage}
}
\description{
Garbage collector for the current modello session.
}
\author{
Filippo MOnari
}
