% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modello.init}
\alias{modello.init}
\title{Initialise Modello Session}
\usage{
modello.init(
  n.numbers = 1e+06,
  n.nodes = 1e+06,
  n.graphs = 1e+06,
  n.opts = 1e+07
)
}
\arguments{
\item{n.numbers}{number of \code{numbers} (size of \code{NUMBERS_})}

\item{n.nodes}{number of \code{nodes} (size of \code{NODES_})}

\item{n.graphs}{number of \code{graphs} (size of \code{GRAPHS_})}

\item{n.opts}{number of \code{optimisers} (size of \code{OPTS_})}
}
\value{
Return invisible NULL
}
\description{
Allocates all the arrays necessary to initialise
the modello session.
}
\examples{
modello.init(10, 10, 10, 10) # Initilaises the modello session with arrays of the given size
modello.is.init() # TRUE
modello.reset()   # restart the modello session with initial array sizes 
modello.is.init() # TRUE
modello.close()   # Closes modello
modello.is.init() # FALSE
}
\author{
Filippo Monari
}
