% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modello.R
\docType{package}
\name{modello}
\alias{modello}
\title{modello: A package for creating computational graphs.}
\description{
the details here
}
\section{Example section}{

The content of the example section here
}

