% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modello.reset}
\alias{modello.reset}
\title{Reset Modello Session}
\usage{
modello.reset()
}
\value{
Return invisible NULL
}
\description{
Rest the modello session, that is
deallocates and reallocates all the arrays.
All the data stored in them is lost.
}
\examples{
modello.init(10, 10, 10, 10) # Initilaises the modello session with arrays of the given size
modello.is.init() # TRUE
modello.reset()   # restart the modello session with initial array sizes 
modello.is.init() # TRUE
modello.close()   # Closes modello
modello.is.init() # FALSE
}
\author{
Filippo Monari
}
