% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module.logistic}
\alias{module.logistic}
\title{R6 class representing a logistic regression model.}
\description{
R6 class representing a logistic regression model.

R6 class representing a logistic regression model.
}
\author{
Filippo Monari
}
\section{Super classes}{
\code{\link[modello:module]{modello::module}} -> \code{\link[modello:module.lm]{modello::module.lm}} -> \code{module.logistic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{module.logistic$new()}}
\item \href{#method-op}{\code{module.logistic$op()}}
\item \href{#method-clone}{\code{module.logistic$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="modello" data-topic="module" data-id="j">}\href{../../modello/html/module.html#method-j}{\code{modello::module$j()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="modello" data-topic="module" data-id="name">}\href{../../modello/html/module.html#method-name}{\code{modello::module$name()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="modello" data-topic="module" data-id="npars">}\href{../../modello/html/module.html#method-npars}{\code{modello::module$npars()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="modello" data-topic="module" data-id="save">}\href{../../modello/html/module.html#method-save}{\code{modello::module$save()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="modello" data-topic="module" data-id="y">}\href{../../modello/html/module.html#method-y}{\code{modello::module$y()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="modello" data-topic="module.lm" data-id="obj">}\href{../../modello/html/module.lm.html#method-obj}{\code{modello::module.lm$obj()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="modello" data-topic="module.lm" data-id="pars">}\href{../../modello/html/module.lm.html#method-pars}{\code{modello::module.lm$pars()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
The initialisation method sets the weights (W), the bais (B) matrices,
the objective function (obj), as well as the name of the module.
The calculation performed is the following
ans = sigmoid(W.op(X) + B)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{module.logistic$new(tx, nin, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tx}}{if TRUE it traspose the input matrix}

\item{\code{nin}}{number of column of the input matrix}

\item{\code{name}}{name of the layer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns the total number of parameters
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-op"></a>}}
\if{latex}{\out{\hypertarget{method-op}{}}}
\subsection{Method \code{op()}}{
Performs: ans = sigmoid(W.op(X) + B)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{module.logistic$op(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{input matrix, reference object of class 'number'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a reference object of class 'number'
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{module.logistic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
