% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_from_RDS}
\alias{module_from_RDS}
\title{Load Module from RDS}
\usage{
module_from_RDS(inf, ...)
}
\arguments{
\item{inf}{input file name}

\item{...}{additional arguments to be passed to \code{readRDS}}
}
\value{
Returns the saved module
}
\description{
Load a module from and RDS file.
}
\author{
Filippo Monari
}
