% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{number_from_list}
\alias{number_from_list}
\title{Number from Collected List}
\usage{
number_from_list(args)
}
\arguments{
\item{args}{list of collected arguments}
}
\value{
Returns the created number invisibly.
}
\description{
Given a list created with a 'number' collection method,
it recreates that 'number'
}
\author{
Filippo Monari
}
