% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{number_from_RDS}
\alias{number_from_RDS}
\title{Load Number from RDS}
\usage{
number_from_RDS(inf, ...)
}
\arguments{
\item{inf}{rds file name/path}

\item{...}{additional arguments for the readRDS function}
}
\value{
Returns a number
}
\description{
Load a number form an RDS file created with the save method.
}
\author{
Filippo
}
