% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pop}
\alias{pop}
\alias{pop.modello_number}
\alias{pop.graph}
\title{Pop Modello Objects}
\usage{
pop(x)

\method{pop}{modello_number}(x)

\method{pop}{graph}(x)
}
\arguments{
\item{x}{a reference object of call 'number' or 'graph'}
}
\value{
Returns invisible x
}
\description{
S3 method to pop (remove) \code{numbers} or
\code{graphs} from their respective arrays
(\code{NUMBERS_} or \code{GRAPHS_}).
}
\examples{
modello.init()
x = number(1)
x$is.linked() # TRUE
pop(x)
x$is.linked() # FALSE
modello.close()
modello.init()
g = graph.open()
graph.close()
g$is.linked() # TRUE
pop(g)
g$is.linked() # FALSE
modello.close()
}
\author{
Filippo Monari
}
