% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sgd.opt}
\alias{sgd.opt}
\title{Create SGD optimiser}
\usage{
sgd.opt(pars)
}
\arguments{
\item{pars}{list of reference object of class 'number' representing
the parameters to optmise.}
}
\value{
a reference object of class 'sgd.opt'
}
\description{
Creates an stochastic gradient descent optimiser.
}
\examples{
modello.init(10, 10, 10, 10)
mdl = module.lm$new(1, 3, 1)
optimizer = sgd.opt(mdl$pars())
is.opt(optimizer) # TRUE
modello.close()
}
