% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sgdwm.opt}
\alias{sgdwm.opt}
\title{Create SGDWM optmiser}
\usage{
sgdwm.opt(pars)
}
\arguments{
\item{pars}{list of reference object of class 'number' representing
the parameters to optmise.}
}
\value{
a reference object of class 'sgdwm.opt'
}
\description{
Creates a stochastic gradient descent with momentum optimiser.
}
\examples{
modello.init()
mdl = module.lm$new(1, 3, 1)
optimizer = sgdwm.opt(mdl$pars())
is.opt(optimizer) # TRUE
modello.close()
}
