% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_math.R
\name{/.modello_number}
\alias{/.modello_number}
\title{Number Division}
\usage{
\method{/}{modello_number}(e1, e2)
}
\arguments{
\item{e1}{a reference object of class 'number'}

\item{e2}{a reference object of class 'number'}
}
\value{
Returns a reference object of class 'number'
}
\description{
Divides a number by another.
}
\examples{
modello.init()
## Scalar / scalar
x1 = number(1, scalar=TRUE)
x2 = number(2, scalar=TRUE)
x3 = x1 / x2
print(x3)
print(x3$v)
## Array / scalar
x4 = number(matrix(2, 2, 2))
x5 = x1 / x4
print(x5)
print(x5$v)
##Array / array
x6 = number(c(2, 1))
x7 = x4 / x6
print(x7)
print(x7$v)
modello.close()
}
\author{
Filippo Monari
}
