% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_math.R
\name{ssq.modello_number}
\alias{ssq.modello_number}
\alias{ssq.default}
\alias{ssq}
\title{Sum of Squares}
\usage{
\method{ssq}{modello_number}(x)

\method{ssq}{default}(x)

ssq(x)
}
\arguments{
\item{x}{a 'numeric' or a reference object of class 'number'}
}
\value{
Returns a 'numeric' or reference object of class 'number'
}
\description{
Sum of square function
\deqn{ssq = \sum_i x_i^2}
}
\examples{
modello.init()
## For modello_numbers
x = number(rnorm(10))
y = ssq(x)
print(y)
print(y$v)
modello.close()
## For numerics
x = rnorm(10)
y = ssq(x)
print(y)
}
\author{
Filippo Monari
}
