% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.modello_number}
\alias{[.modello_number}
\title{Silce Numbers}
\usage{
\method{[}{modello_number}(x, ..., contiguous = TRUE)
}
\arguments{
\item{x}{reference object of class 'number'}

\item{...}{additional parametrers to pass to the slice methods}

\item{contiguous}{if TRUE it tryes to create a contiguous slice}
}
\value{
Returns a reference object of class 'number'
}
\description{
Slices \code{numbers}
}
\examples{
modello.init(10, 10, 10, 10)
x = number(matrix(rnorm(9), 3, 3))
print(x$v)
s1 = x[1, 2]
print(s1)
print(s1$v)
s2 = x[1,2, contiguous=FALSE]
print(s2)
print(s2$v)
modello.close()
}
\author{
Filippo Monari
}
