% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[<-.modello_number}
\alias{[<-.modello_number}
\title{Sets the value of a slice in a \code{number}}
\usage{
\method{[}{modello_number}(x, ..., dx = FALSE) <- value
}
\arguments{
\item{x}{\code{number}}

\item{...}{indexes defining the slice}

\item{dx}{if TRUE the slice is taken from the gradient}

\item{value}{value to set}
}
\description{
Sets the value of a slice in a \code{number}
}
\examples{
modello.init(10, 10, 10, 10)
x = number(matrix(rnorm(9), 3, 3))
print(x$v)
x[1,1] = -100
print(x$v)
x[1:2,1:2] = 100
print(x$v)
modello.close()
}
