% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers_math.R
\name{sum.modello_number}
\alias{sum.modello_number}
\title{Sum}
\usage{
\method{sum}{modello_number}(x, k = 0, ...)
}
\arguments{
\item{x}{a reference object of class 'number'}

\item{k}{integer, dimension along wihich do the reduction. Currently constrained to be <= 2}

\item{...}{numerics to sum. Ignored for \code{modello_numbers}}
}
\value{
Returns a 'numeric' or reference object of class 'number'
}
\description{
Sum function
}
\examples{
modello.init(10, 10, 10, 10)
x = number(rep(1, 20))
y = sum(x)
print(y)
print(y$v)
modello.close()
}
\author{
Filippo Monari
}
